Issue #62: Refactor withXml().
