+++ src/Messages/Components/Identifier.php
@@ -269,7 +269,7 @@ class Identifier extends AbstractComponent
     /**
      * {@inheritdoc}
      */
-    protected function parseXml($xml)
+    public function withXml($xml)
     {
         $parser = $this->getParser();
         $parser->addXmlContent($xml);
