+++ src/Messages/Components/ReferenceDocument.php
@@ -248,7 +248,7 @@ class ReferenceDocument extends AbstractComponent
     /**
      * {@inheritdoc}
      */
-    protected function parseXml($xml)
+    public function withXml($xml)
     {
         $parser = $this->getParser();
         $parser->addXmlContent($xml);
