+++ src/Messages/Components/ReturnAddress.php
@@ -202,7 +202,7 @@ class ReturnAddress extends AbstractComponent
     /**
      * {@inheritdoc}
      */
-    protected function parseXml($xml)
+    public function withXml($xml)
     {
         $parser = $this->getParser();
         $parser->addXmlContent($xml);
