+++ src/Messages/Components/SourceLanguage.php
@@ -80,7 +80,7 @@ class SourceLanguage extends AbstractComponent
     /**
      * {@inheritdoc}
      */
-    protected function parseXml($xml)
+    public function withXml($xml)
     {
         $parser = $this->getParser();
         $parser->addXmlContent($xml);
