+++ src/Messages/Components/Status.php
@@ -200,7 +200,7 @@ class Status extends AbstractComponent
     /**
      * {@inheritdoc}
      */
-    protected function parseXml($xml)
+    public function withXml($xml)
     {
         $parser = $this->getParser();
         $parser->addXmlContent($xml);
