+++ src/Messages/Components/Target.php
@@ -306,7 +306,7 @@ class Target extends AbstractComponent
     /**
      * {@inheritdoc}
      */
-    protected function parseXml($xml)
+    public function withXml($xml)
     {
         $parser = $this->getParser();
         $parser->addXmlContent($xml);
