+++ src/Messages/Notifications/AbstractNotification.php
@@ -14,7 +14,6 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
  */
 abstract class AbstractNotification extends AbstractMessage implements ParserAwareInterface, EventSubscriberInterface
 {
-
     /**
      * {@inheritdoc}
      */
@@ -31,32 +30,4 @@ abstract class AbstractNotification extends AbstractMessage implements ParserAwa
      * @return mixed
      */
     abstract public function onParseNotification(ParseNotificationEvent $event);
-
-    /**
-     * Set a message or a component internal properties given its XML representation.
-     *
-     * @param string $xml
-     *      XML string.
-     *
-     * @return \EC\Poetry\Messages\MessageInterface|\EC\Poetry\Messages\ComponentInterface
-     */
-    public function withXml($xml)
-    {
-        $this->setRaw($xml);
-
-        return $this->parseXml($xml);
-    }
-
-    /**
-     * Parse a XML string into a set of properties.
-     *
-     * @param string $xml
-     *      XML string.
-     *
-     * @return \EC\Poetry\Messages\MessageInterface|\EC\Poetry\Messages\ComponentInterface
-     */
-    protected function parseXml($xml)
-    {
-        return $this;
-    }
 }
