+++ src/Messages/Notifications/StatusUpdated.php
@@ -43,15 +43,12 @@ class StatusUpdated extends AbstractNotification
     /**
      * {@inheritdoc}
      */
-    protected function parseXml($xml)
+    public function withXml($xml)
     {
+        parent::withXml($xml);
+
         $parser = $this->getParser();
         $parser->addXmlContent($xml);
-
-        $xml = $parser->getOuterContent('POETRY/request/demandeId');
-        $this->getIdentifier()->withXml($xml);
-
-        $this->setMessageId($parser->getAttribute('POETRY/request', 'id'));
 
         $parser->eachComponent("POETRY/request/status", function (Parser $component) {
             $this->withStatus()
