+++ src/Messages/Requests/AbstractRequest.php
@@ -52,32 +52,4 @@ abstract class AbstractRequest extends AbstractMessage implements ParserAwareInt
      * @return string
      */
     abstract public function getType();
-
-    /**
-     * Set a message or a component internal properties given its XML representation.
-     *
-     * @param string $xml
-     *      XML string.
-     *
-     * @return \EC\Poetry\Messages\MessageInterface|\EC\Poetry\Messages\ComponentInterface
-     */
-    public function withXml($xml)
-    {
-        $this->setRaw($xml);
-
-        return $this->parseXml($xml);
-    }
-
-    /**
-     * Parse a XML string into a set of properties.
-     *
-     * @param string $xml
-     *      XML string.
-     *
-     * @return \EC\Poetry\Messages\MessageInterface|\EC\Poetry\Messages\ComponentInterface
-     */
-    protected function parseXml($xml)
-    {
-        return $this;
-    }
 }
